third_party/WebKit/LayoutTests/media/autoplay/webaudio-autoplay-no-gesture.html

<!DOCTYPE html>
<title>Test that webaudio can be created without a user gesture</title>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script>
test(t => {
  // Make sure this overrides any preset autoplay policy.
  internals.settings.setAutoplayPolicy('document-user-activation-required');

  try {
    const ac = new AudioContext();
    assert_equals('running', ac.state);
  } catch (e) {
    assert_unreached();
  }
});
</script>
